Split survey management controllers into modules
